+++ golden/components/App.html
@@ -2,14 +2,3 @@
     <Nav />
     <div>Brian Was Here</div>
 </main>
-
-<style>
-    main {
-        position: relative;
-        max-width: 56em;
-        background-color: white;
-        padding: 2em;
-        margin: 0 auto;
-        box-sizing: border-box;
-    }
-</style>
