+++ golden/components/Todo.html
@@ -1,6 +1,6 @@
 <!-- maybe factor can somehow "bind" the models.Todo to this html? -->
 <div>
-    <h1>{{ .Name }}</h1>
-    <small>{{ .Description }}</small>
-    <div>(<a href="{{ .Permalink }}">Permalink</a>)</div>
+    <h1>vecty-field:Name</h1>
+    <small>vecty-field:Description</small>
+    <div>(<a href="vecty-field:Permalink">Permalink</a>)</div>
 </div>
