+++ golden/components/markdown.html
@@ -1,5 +1,5 @@
 <body>
     <div style="float: right;">
-        <textarea style="font-family: monospace;" cols="70" rows="14" oninput="texthandler">{{ p.Input }}</textarea>
+        <textarea style="font-family: monospace;" cols="70" rows="14" oninput="texthandler">vecty-data:Input</textarea>
     </div>
 </body>
