+++ golden/routes/error_generated.go
@@ -1,44 +1,29 @@
-package routes
+// This file was created with https://jsgo.io/dave/html2vecty
+package main
 
 import (
-	"github.com/bketelsen/factor/markup"
+	"github.com/gowasm/vecty"
+	"github.com/gowasm/vecty/elem"
 )
 
-type Error struct {
+func main() {
+	vecty.RenderBody(&Page{})
 }
 
-var ErrorTemplate = `<h1>status</h1>
-
-<p>error.message</p>`
-var ErrorStyles = `
-	h1,
-	p {
-		margin: 0 auto;
-	}
-
-	h1 {
-		font-size: 2.8em;
-		font-weight: 700;
-		margin: 0 0 0.5em 0;
-	}
-
-	p {
-		margin: 1em auto;
-	}
-
-	@media (min-width: 480px) {
-		h1 {
-			font-size: 4em;
-		}
-	}
-`
-
-func (t *Error) Render() string {
-	return ErrorTemplate
+type Page struct {
+	vecty.Core
 }
-func (t *Error) Style() string {
-	return ErrorStyles
-}
-func init() {
-	markup.Register(&Error{})
+
+func (p *Page) Render() vecty.ComponentOrHTML {
+	return elem.Body(
+		elem.Heading1(
+			vecty.Text("status"),
+		),
+		elem.Paragraph(
+			vecty.Text("error.message"),
+		),
+		elem.Style(
+			vecty.Text("h1,\n\tp {\n\t\tmargin: 0 auto;\n\t}\n\n\th1 {\n\t\tfont-size: 2.8em;\n\t\tfont-weight: 700;\n\t\tmargin: 0 0 0.5em 0;\n\t}\n\n\tp {\n\t\tmargin: 1em auto;\n\t}\n\n\t@media (min-width: 480px) {\n\t\th1 {\n\t\t\tfont-size: 4em;\n\t\t}\n\t}"),
+		),
+	)
 }
