+++ golden/routes/todoid_generated.go
@@ -1,23 +1,26 @@
-package routes
+// This file was created with https://jsgo.io/dave/html2vecty
+package main
 
 import (
-	"github.com/bketelsen/factor/markup"
+	"github.com/gowasm/vecty"
+	"github.com/gowasm/vecty/elem"
 )
 
-type Todoid struct {
+func main() {
+	vecty.RenderBody(&Page{})
 }
 
-var TodoidTemplate = `<h1>Todo</h1>
-
-<p>I'm not sure exactly how this'll work yet. Haven't run factor dev against it yet...</p>`
-var TodoidStyles = ``
-
-func (t *Todoid) Render() string {
-	return TodoidTemplate
+type Page struct {
+	vecty.Core
 }
-func (t *Todoid) Style() string {
-	return TodoidStyles
-}
-func init() {
-	markup.Register(&Todoid{})
+
+func (p *Page) Render() vecty.ComponentOrHTML {
+	return elem.Body(
+		elem.Heading1(
+			vecty.Text("Todo"),
+		),
+		elem.Paragraph(
+			vecty.Text("I'm not sure exactly how this'll work yet. Haven't run factor dev against it yet..."),
+		),
+	)
 }
