+++ golden/routes/todos_generated.go
@@ -1,51 +1,27 @@
-package routes
+// This file was created with https://jsgo.io/dave/html2vecty
+package main
 
 import (
-	"fmt"
-
-	"github.com/bketelsen/factor/golden/models"
-	"github.com/bketelsen/factor/markup"
-	"github.com/satori/go.uuid"
+	"github.com/gowasm/vecty"
+	"github.com/gowasm/vecty/elem"
 )
 
-type Todos struct {
-	List []*models.Todo
+func main() {
+	vecty.RenderBody(&Page{})
 }
 
-//    <Todo Name="{{.Name}}" Description="{{.Description}}" Permalink="{{.Permalink}}"/>
-var TodosTemplate = `<div class="todos">
-{{ range .List }}
-<Todo Name="{{.Name}}" Description="{{.Description}}" Permalink="{{.Permalink}}"/>
-{{ end }}
-</div>`
-var TodosStyles = ``
-
-func (t *Todos) Render() string {
-	tdc := new(models.TodoClient)
-	uid := uuid.Must(uuid.NewV4())
-	fmt.Println(uid.String())
-	todo, err := tdc.Get(uid)
-	if err != nil {
-		fmt.Println(err)
-		panic(err)
-	}
-
-	/*	todo := &models.Todo{
-			Name:        "Brian",
-			Description: "Des",
-			Permalink:   "/1",
-		}
-	*/
-	t.List = []*models.Todo{todo}
-	return TodosTemplate
-}
-func (t *Todos) Style() string {
-	return TodosStyles
-}
-func (t *Todos) OnMount() {
-
+type Page struct {
+	vecty.Core
 }
 
-func init() {
-	markup.Register(&Todos{})
+func (p *Page) Render() vecty.ComponentOrHTML {
+	return elem.Body(
+		elem.Div(
+			vecty.Markup(
+				vecty.Class("todos"),
+			),
+			vecty.Text("{{ range .Todos }}"),
+			vecty.Text("{{ .TodoComponent . }}\n    {{ end }}"),
+		),
+	)
 }
